Remove main menu and AFK popup hacks
As we're moving towards custom popups with XML instead of StaticPopup, we no longer need to work around certain interactions that required hacks implemented in the addon.

One of the more annoying examples was the main menu firing a popup on logout, which even in rested zones would interfere if a Sippy Cup StaticPopup was visible. So we had to hide those in the past.

Same with the ElvUI AFK screen — it changed what was on the screen and would inadvertently break the StaticPopups that were shown.

Both these hacks/workarounds are no longer required, so remove them from the main SippyCup.lua file. The actual removal of the code will be in one of the follow-up commits.