Rewrite HandleAlerts
Whether an alert can play or not (and the throttling for that) is now handled by the popup code (in a later commit), so that is no longer necessary to be handled within the function itself.

The only thing that still required handling was quantity throttling to make sure that when a lot of popups pop at once, think like on startup, that the user doesn't get ear attack-- serenaded with a whole orchestra of sound effects all at once.