Rewrite nontrackable to no aura (split itemTrackable & spellTrackable)
Some consumables are special because they do not set an aura on the player so we cannot track them like usual.

For no aura consumables there are two variants. The simpler one is spellTrackable. They have an aura and a cooldown you can track reliably.

The more complex variant is itemTrackable like the Decorative Yard Flamingo which can only be tracked via the item's cooldown. This method is less reliable.

The code is rewritten to support both types. Some desyncs or issues may remain since testing was done on PTR and not on heavy load servers.