Item loading only at startup, not everywhere, to prevent weird issues
There were situations that the settings would not show some items due to consistently trying to gather data of items that might not be around on X or Y moment.

We just take care of it on startup, and 'hope' for the best.

There is some extra code in there for deferSetup, but we don't utilize that just yet (or will ever).