Tweak popup data handling
This is a bit of a bigger commit.

There's now a data table sent towards the popup queue, givinng some information it might need to decide whether a popup should be fired or not.

This changes how things are done considerably, but the user outcome should be the same (or even better). This outcome also allows us to extend the data sent through as well in an easier way.