Handle combat & loading screen deferring
This extends HandleDeferredActions to handle combat- and loading screen-related deferring instead of just bag-related.

This, hopefully, will help with queueing up reminder popups that happen during the loading screen or during combat at a time it is more permissable and sensible for Sippy Cup to do so.

The combat testing was not extensive, and issues could still be here.