Add delayedAura consumable parameter
Some consumables only apply their RP effect through a delayed (food) buff. This means the actual `UNIT_AURA` fire is delayed.

This makes the popup stay there with a disabled refresh button for the x amount of seconds until the buff is actually applied.

Which confused some users, so now we have a disable reason that tells users why exactly they keep seeing the popup with the button disabled.