Initial toys support
A lot of this commit is adding in the back-end to track toys, which is somewhat simplified given that a lot of that had already been done to track certain consumables.

A lot of things also got renamed from using "consumable" to "option", as that's what I call them both: options.

Due to how some toys work, I had to add in a cooldown mismatch option (with a custom tga). Plus some re-coding of updating aurainstanceid's which for example the Uncracked Cold Ones suffered from.

This is not ready for release, but does not seem to have major issues in my testing. It will require a few more (optimization) passes.

Translations will be off for a bit, working on that as well.