Handle event registration better to prevent bad login/reloads.
This was one of those niche issues that only happened every so often, however when it did Sippy Cup would break entirely.

It basically pretended it worked, but given it was missing some very necessary registrations that were meant to fire on login, it was not able to set itself up correctly.

Moving these around to `PLAYER_LOGIN` to register, even if DB and options are not necessarily loaded does the trick and should not break as it doesn't do anything it shouldn't wrong realistically.